refactor(App): separate App container and stateless component
